Refactor the files not needed now, maybe used in the future if we have time enough to implement it. Writeups are now stored in the category they belong to, and when a writeup is created, it tells you were it's being stored.
