+++ src/main/resources/templates/creation_pages/new_writeup.html
@@ -11,24 +11,24 @@
     <p>Title:</p>
     <input type='text' name='title'/>
     <p>Text:</p>
-    <h3>Category: </h3>
     <textarea name='text' rows=5 cols=40></textarea>
-    <input type="radio" id="crypto" name="topic" value="crypto">
+    <h3>Category: </h3>
+    <input type="radio" id="crypto" name="Category" value="crypto" required>
     <label for="crypto">Crypto</label><br>
 
-    <input type="radio" id="pwn" name="topic" value="pwn">
+    <input type="radio" id="pwn" name="Category" value="pwn">
     <label for="pwn">Pwn</label><br>
 
-    <input type="radio" id="reversing" name="topic" value="reversing">
+    <input type="radio" id="reversing" name="Category" value="reversing" required>
     <label for="reversing">Reversing</label><br>
 
-    <input type="radio" id="web" name="topic" value="web">
+    <input type="radio" id="web" name="Category" value="web" required>
     <label for="web">Web</label><br>
 
-    <input type="radio" id="mobile" name="topic" value="mobile">
+    <input type="radio" id="mobile" name="Category" value="mobile" required>
     <label for="mobile">Mobile</label><br>
 
-    <input type="radio" id="osint" name="topic" value="osint">
+    <input type="radio" id="osint" name="Category" value="osint" required>
     <label for="osint">OSINT</label><br>
 
     <p>Image file:</p>
