+++ src/main/resources/templates/for_future_use/new_post.html
