+++ src/main/resources/templates/saved_posts.html
@@ -2,6 +2,9 @@
 <body>
 <p>Post has been saved.</p>
 <p>You have created {{numPosts}} so far.</p>
+{{#storageLocation}}
+    <p>The Writeup was stored in {{storageLocation}}</p>
+{{/storageLocation}}
 <a href="/">Back</a>
 </body>
 </html>
